fix: resolve remaining CodeQL alerts in production code
- DateFunctions: cast int to double before multiplication to prevent precision loss
- CachedMetadataProvider: document double-check-lock pattern for CodeQL
- EnvironmentConfigService: document nullable coalescing reachability
- EnvironmentSelectorDialog: extract complex condition to named booleans
- EnvironmentConfigDialog: add missing XML summary on constructor
- TuiDialog: suppress virtual-call-in-constructor (Terminal.Gui pattern)
- DataTableView: suppress virtual-call-in-constructor (Terminal.Gui pattern)
- DataverseConnectionPool: document intentional static field write under lock
- UsernamePasswordCredentialProvider: document intentional obsolete API usage

+++ src/PPDS.Dataverse/Query/Execution/Functions/DateFunctions.cs
@@ -183,7 +183,7 @@ public static class DateFunctions
                 "month" => dt.Value.AddMonths(number),
                 "day" => dt.Value.AddDays(number),
                 "dayofyear" => dt.Value.AddDays(number),
-                "week" => dt.Value.AddDays(number * 7),
+                "week" => dt.Value.AddDays((double)number * 7),
                 "hour" => dt.Value.AddHours(number),
                 "minute" => dt.Value.AddMinutes(number),
                 "second" => dt.Value.AddSeconds(number),
